+++ .storybook/preview.js
@@ -1,9 +1,16 @@
+import i18next from '../src/i18n'
+
 export const parameters = {
-  actions: { argTypesRegex: "^on[A-Z].*" },
+  actions: { argTypesRegex: '^on[A-Z].*' },
   controls: {
     matchers: {
       color: /(background|color)$/i,
-      date: /Date$/,
-    },
+      date: /Date$/
+    }
   },
+  i18next,
+  locale: 'en',
+  locales: {
+    en: 'English'
+  }
 }
