+++ src/stories/hello-dataverse/HelloDataverse.stories.tsx
@@ -1,17 +1,17 @@
 import { ComponentStory, ComponentMeta } from '@storybook/react'
 import { within, userEvent } from '@storybook/testing-library'
-import { Page } from './Page'
+import { HelloDataverse } from '../../sections/hello-dataverse/HelloDataverse'
 
 export default {
   title: 'Hello Dataverse/Page',
-  component: Page,
+  component: HelloDataverse,
   parameters: {
     // More on Story layout: https://storybook.js.org/docs/react/configure/story-layout
     layout: 'fullscreen'
   }
-} as ComponentMeta<typeof Page>
+} as ComponentMeta<typeof HelloDataverse>
 
-const Template: ComponentStory<typeof Page> = (args) => <Page {...args} />
+const Template: ComponentStory<typeof HelloDataverse> = (args) => <HelloDataverse />
 
 export const LoggedOut = Template.bind({})
